update route, tests, refactoring

+++ StockApi.WebApi.NetCore/StockApi/Models/StockItemDto.cs
@@ -29,7 +29,7 @@
 			}
 
 			this.Id = stockItem.Id;
-			this.InStock = (int) stockItem.InStock;
+			this.InStock = stockItem.InStock;
 		}
 
 		/// <summary>
